Changes the project to be a pug file served via Flask, from being a HTML file accessed directly. This necessitated a directory structure and everything!
You can configure the cameras and the server from the config folder.

I've also figured out a bunch of things, like:
* Pug + Flasks means you need to be careful of curly braces
* The ?time= argument is just a dummy to trick the browser into
  reloading the image.
