+++ camera-hacker/Config/Cameras.py
@@ -1,0 +1,3 @@
+IPS = [
+    'http://192.168.1.3:8080'
+]
