+++ camera-hacker/Config/Server.py
@@ -1,0 +1,4 @@
+HOST = 'localhost'  # change this to 0.0.0.0 for production; it ruins your ability to watch,
+                    # but is visible beyond your local loopback
+
+PORT = 8080  # remember that port 80 requires super user permissions on *nix
