+++ camera-hacker/templates/feed.pug
@@ -1,0 +1,15 @@
+html
+    head
+        script(src="{{url_for('static', filename='videoScript.js')}}" type="text/javascript")
+    body
+        div(style='display:flex;flex-direction:column;flex-wrap:wrap;justify-content:flex-start')
+            each ip, ix in cameras
+                p
+                    // PLEASE FOR THE LOVE OF GOD DON'T MOVE THESE ONTO THE SAME LINE
+                        THE UNHOLY GODS FOR THE UNHOLY STACK OF TECHNOLOGY I'M USING TO MAKE THIS
+                        ALL WORK DO _NOT_ APPRECIATE IT. FROM TRIAL AND ERROR, I CAN SAY THAT
+                        EVERYTHING GOES TO SHIT AS SOON AS IT DETECTS TWO CURLY BRACES ON THE SAME
+                        LINE >:(
+                    img(id="cameraImage#{ix}"
+                        src="#{ip}/camera.jpg"
+                        onload="loadImage('cameraImage#{ix}')")
